fix linting and remove redundant files

+++ src/geodesic-grid/gauss_legendre.cpp
@@ -33,7 +33,6 @@ GaussLegendreGrid::GaussLegendreGrid(MeshBlockPack *pmy_pack, int ntheta, Real r
   interp_indcs("interp_indcs",1,1),
   interp_wghts("interp_wghts",1,1,1),
   interp_vals("interp_vals",1) {
-  
   // reallocate and set interpolation coordinates, indices, and weights
   int &ng = pmy_pack->pmesh->mb_indcs.ng;
   nangles = 2*ntheta*ntheta;
@@ -68,7 +67,7 @@ void GaussLegendreGrid::InitializeAngleAndWeights() {
     // calculate theta
     polar_pos.h_view(n,0) = acos(roots_and_weights[0][n%ntheta]);
     // calculate phi
-    polar_pos.h_view(n,1) = 2*M_PI/(2*ntheta)*int(n/ntheta);
+    polar_pos.h_view(n,1) = 2*M_PI/(2*ntheta)*(n/ntheta);
   }
 
   // sync to device
